:truck: Move images to static folder
