+++ src/components/wedding/About/index.tsx
@@ -1,18 +1,18 @@
 import FlipCard from './FlipCard';
 import HorizontalRule from '../HorizontalRule';
 
-import Casper from '../../../images/Casper.webp';
-import Ella from '../../../images/Ella.webp';
-import Jeb from '../../../images/Jeb.webp';
-import Jeremy from '../../../images/Jeremy.webp';
-import Jesse from '../../../images/Jesse.webp';
-import Jordan from '../../../images/Jordan.webp';
-import Kelly from '../../../images/Kelly.webp';
-import Kristopher from '../../../images/Kristopher.webp';
-import Nick from '../../../images/Nick.webp';
-import Sarah from '../../../images/Sarah.webp';
-import Shannon from '../../../images/Shannon.webp';
-import Will from '../../../images/Will.webp';
+import Casper from '../../../../static/images/Casper.webp';
+import Ella from '../../../../static/images/Ella.webp';
+import Jeb from '../../../../static/images/Jeb.webp';
+import Jeremy from '../../../../static/images/Jeremy.webp';
+import Jesse from '../../../../static/images/Jesse.webp';
+import Jordan from '../../../../static/images/Jordan.webp';
+import Kelly from '../../../../static/images/Kelly.webp';
+import Kristopher from '../../../../static/images/Kristopher.webp';
+import Nick from '../../../../static/images/Nick.webp';
+import Sarah from '../../../../static/images/Sarah.webp';
+import Shannon from '../../../../static/images/Shannon.webp';
+import Will from '../../../../static/images/Will.webp';
 
 import styles from './About.module.css';
